requirements.txt
attrs==19.3.0
importlib-metadata==1.4.0
more-itertools==8.1.0
numpy==1.18.1
packaging==20.0
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.4
six==1.14.0
wcwidth==0.1.8
zipp==2.0.0
